+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the "Total Recall" extension will be documented in this file.
 
+## [0.6.1]
+
+- Bug fixes
+- Better code structure
+
+## [0.5.7]
+
+- Fixed a bug where the extension would close all editors when switching to a new branch
+- Setup prettier
+- Better code structure
+
 ## [0.5.6]
 
 - Updated README